pyproject.toml
[project]
name = "zsuite"
version = "0.8.1"
description = "Lightweight Python utility library"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
dependencies = [
    "tzdata",
    "cryptography",
    "python-dateutil",
    "python-dotenv",
    "PyYAML",
]

[dependency-groups]
dev = [
    "pytest",
    "ruff>=0.14.0",
    "bump-my-version",
]

# ---- Build System ----
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
exclude = ["tests*"]

# ---- Ruff - Linter & Formatter ----
[tool.ruff]
target-version = "py310"
line-length = 120
indent-width = 4
respect-gitignore = true
show-fixes = true

# ---- Ruff Linting Rules ----
[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "UP",     # pyupgrade
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "SIM",    # flake8-simplify
    "RUF",    # Ruff-specific rules
]

ignore = [
    "PLR2004",  # Magic value in comparison
    "BLE001",   # Bare except allowed
]

fixable = ["ALL"]
extend-fixable = ["B"]
unfixable = [
    "T201", # Never auto-remove print statements
]

# ---- Ruff Per-File Ignores ----
[tool.ruff.lint.per-file-ignores]
"tests/*" = [
    "E501", "S101", "D", "ANN", "ARG001", "ARG002",
    "SLF001", "T201",
]
"__init__.py" = ["F401"]

# ---- Ruff Formatter ----
[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
docstring-code-format = true
docstring-code-line-length = "dynamic"

# ---- Pytest Configuration ----
[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

# ---- Version Bumping ----
[tool.bumpversion]
current_version = "0.8.1"
parse = "(?P<major>\\d+)\\.(?P<minor>\\d+)\\.(?P<patch>\\d+)"
serialize = ["{major}.{minor}.{patch}"]
search = "{current_version}"
replace = "{new_version}"
regex = false
ignore_missing_version = false
tag = true
sign_tags = false
tag_name = "{new_version}"
tag_message = "Bump version: {current_version} → {new_version}"
allow_dirty = false
commit = true
message = "Bump version: {current_version} → {new_version}"
commit_args = ""
pre_commit_hooks = ["uv sync", "git add uv.lock"]

[[tool.bumpversion.files]]
filename = "pyproject.toml"
search = 'version = "{current_version}"'
replace = 'version = "{new_version}"'
